test fichier JSrequest lien fonctionnels

+++ models/jsRequest/deleteEvent.php
@@ -1,17 +1,21 @@
 <?php
-require_once('../connexionManager.php');
+if (isset($_POST['eventId']))
+{
+    require_once('../connexionManager.php');
 
-$dbConnection = new ConnexionManager();
-$db = $dbConnection->dbConnection();
+    $dbConnection = new ConnexionManager();
+    $db = $dbConnection->dbConnection();
 
-$eventId = $_POST['eventId'];
-// Suppression de la table events
-$deleteEvent = $db->prepare('DELETE FROM events WHERE eventId = :eventId');
-$deleteEvent->execute(array(
-    'eventId' => $eventId
-));
-// Suppression de la table jointable
-$deleteEvent = $db->prepare('DELETE FROM jointable WHERE event_id = :eventId');
-$deleteEvent->execute(array(
-    'eventId' => $eventId
-));
+    $eventId = strip_tags($_POST['eventId']);
+    
+    // Suppression de la table events
+    $deleteEvent = $db->prepare('DELETE FROM events WHERE eventId = :eventId');
+    $deleteEvent->execute(array(
+        'eventId' => $eventId
+    ));
+    // Suppression de la table jointable
+    $deleteEvent = $db->prepare('DELETE FROM jointable WHERE event_id = :eventId');
+    $deleteEvent->execute(array(
+        'eventId' => $eventId
+    ));
+}
